fix favourite button
